Bugfixes for CheckIntFilesIntegrity
Index creation logic for the IntFielsTable was incorrect. Fixed!
NOTE: If this function has been run before you *must* right click the "IntFilesTable" in the Studio's "Table Explorer", select "Remove Table" and then - and this is very important - select the "Delete Table". This will both remove the file from filelist.cfg, remove it from disk *and* remove the SQL table.